Moved request-related utils to their own package. Created test class for ApiKeyRequestDecorator
