+++ app/src/main/java/com/example/campusguide/utils/request/RequestDispatcher.kt
@@ -1,4 +1,4 @@
-package com.example.campusguide.utils
+package com.example.campusguide.utils.request
 
 import org.json.JSONObject
