+++ app/src/test/java/com/example/campusguide/utils/request/RequestQueueTest.kt
@@ -1,4 +1,4 @@
-package com.example.campusguide.utils
+package com.example.campusguide.utils.request
 
 import android.content.Context
 import org.junit.Before
